fix bug consent to lifebank
