+++ hapi/src/api/account.api.js
@@ -15,7 +15,6 @@ const vaultApi = require('./vault.api')
 const preRegLifebank = require('./pre-register.api')
 const verificationCodeApi = require('./verification-code.api')
 const mailApi = require('../utils/mail')
-const lifebankApi = require('./lifebank.api')
 const LIFEBANKCODE_CONTRACT = eosConfig.lifebankCodeContractName
 const MAIL_APPROVE_LIFEBANNK = eosConfig.mailApproveLifebank
 
@@ -309,6 +308,30 @@ const grantConsent = async account => {
   return consentTransaction
 }
 
+const formatSchedule = (schedule) => {
+  let scheduleFormat = ''
+
+  let hours
+  for (hours of schedule)
+    scheduleFormat += `, ${hours.day} ${hours.open} - ${hours.close}`
+
+  return scheduleFormat.replace(',', ' ')
+}
+
+const formatLifebankData = (lifebankData) => {
+  lifebankData.schedule = formatSchedule(JSON.parse(lifebankData.schedule))
+  lifebankData.coordinates = JSON.parse(lifebankData.coordinates)
+  if (lifebankData.immunity_test) lifebankData.immunity_test = 'Yes'
+  else lifebankData.immunity_test = 'No'
+
+  if (lifebankData.urgency_level === 1) lifebankData.urgency_level = 'Low'
+  else if (lifebankData.urgency_level === 2)
+    lifebankData.urgency_level = 'Medium'
+  else lifebankData.urgency_level = 'High'
+
+  return lifebankData
+}
+
 const verifyEmail = async ({ code }) => {
   const resUser = await userApi.verifyEmail({
     verification_code: { _eq: code }
@@ -323,7 +346,7 @@ const verifyEmail = async ({ code }) => {
     resLifebank.update_preregister_lifebank.affected_rows !== 0
   ) {
     if (resLifebank.update_preregister_lifebank.affected_rows !== 0) {
-      resLifebank.update_preregister_lifebank.returning[0] = lifebankApi.formatLifebankData(
+      resLifebank.update_preregister_lifebank.returning[0] = formatLifebankData(
         resLifebank.update_preregister_lifebank.returning[0]
       )
       try {
